+++ lib/broadside/deploy.rb
@@ -1,21 +1,24 @@
 module Broadside
   class Deploy
     include Utils
+    include VerifyInstanceVariables
 
-    attr_accessor :deploy_config
+    attr_reader(
+      :command,
+      :instance,
+      :lines,
+      :tag,
+      :target
+    )
 
-    def initialize(opts)
-      @deploy_config = Broadside.config.deploy.dup
-      @deploy_config.tag = opts[:tag]           || @deploy_config.tag
-      @deploy_config.target = opts[:target]     || @deploy_config.target
-      @deploy_config.verify(:target, :targets)
-      @deploy_config.load_target!
-
-      @deploy_config.scale = opts[:scale]       || @deploy_config.scale
-      @deploy_config.rollback = opts[:rollback] || @deploy_config.rollback
-      @deploy_config.instance = opts[:instance] || @deploy_config.instance
-      @deploy_config.command = opts[:command]   || @deploy_config.command
-      @deploy_config.lines = opts[:lines]       || @deploy_config.lines
+    def initialize(target, opts = {})
+      @target   = target
+      @command  = opts[:command]  || @target.command
+      @instance = opts[:instance] || @target.instance
+      @lines    = opts[:lines]    || 10
+      @rollback = opts[:rollback] || 1
+      @scale    = opts[:scale]    || @target.scale
+      @tag      = opts[:tag]
     end
 
     def short
@@ -23,42 +26,42 @@ module Broadside
     end
 
     def full
-      run_predeploy
+      config.verify(:ssh)
+      verify(:tag)
+
+      info "Running predeploy commands for #{family}..."
+      run_commands(@target.predeploy_commands)
+      info 'Predeploy complete.'
+
       deploy
     end
 
     def deploy
-      @deploy_config.verify(:tag)
+      verify(:tag)
+
       info "Deploying #{image_tag} to #{family}..."
       yield
       info 'Deployment complete.'
     end
 
-    def rollback(count = @deploy_config.rollback)
-      @deploy_config.verify(:rollback)
-      info "Rolling back #{@deploy_config.rollback} release for #{family}..."
+    def rollback(count = @rollback)
+      info "Rolling back #{count} release for #{family}..."
       yield
       info 'Rollback complete.'
     end
 
     def scale
-      info "Rescaling #{family} with scale=#{@deploy_config.scale}"
+      info "Rescaling #{family} with scale=#{@scale}"
       yield
       info 'Rescaling complete.'
     end
 
     def run
-      @deploy_config.verify(:tag, :ssh, :command)
-      info "Running command [#{@deploy_config.command}] for #{family}..."
+      config.verify(:ssh)
+      verify(:tag, :command)
+      info "Running command [#{@command}] for #{family}..."
       yield
       info 'Complete.'
-    end
-
-    def run_predeploy
-      @deploy_config.verify(:tag, :ssh)
-      info "Running predeploy commands for #{family}..."
-      yield
-      info 'Predeploy complete.'
     end
 
     def status
@@ -68,36 +71,33 @@ module Broadside
     end
 
     def logtail
-      @deploy_config.verify(:instance)
+      verify(:instance)
       yield
     end
 
     def ssh
-      @deploy_config.verify(:instance)
+      verify(:instance)
       yield
     end
 
     def bash
-      @deploy_config.verify(:instance)
+      verify(:instance)
       yield
     end
 
-    def container_definitions
-      yield
-    end
-
-    protected
+    private
 
     def family
-      "#{config.base.application}_#{@deploy_config.target}"
+      "#{config.application}_#{@target.name}"
     end
 
     def image_tag
-      "#{config.base.docker_image}:#{@deploy_config.tag}"
+      raise ArgumentError, "Missing tag" unless @tag
+      "#{config.docker_image}:#{@tag}"
     end
 
     def gen_ssh_cmd(ip, options = { tty: false })
-      opts = @deploy_config.ssh || {}
+      opts = config.ssh || {}
       cmd = 'ssh -o StrictHostKeyChecking=no'
       cmd << ' -t -t' if options[:tty]
       cmd << " -i #{opts[:keyfile]}" if opts[:keyfile]