+++ lib/broadside/ecs/ecs_deploy.rb
@@ -12,14 +12,14 @@ module Broadside
       memory: 1000
     }
 
-    def initialize(opts)
-      super(opts)
+    def initialize(target, opts = {})
+      super
       config.ecs.verify(:cluster, :poll_frequency)
     end
 
     def deploy
       super do
-        unless EcsManager.service_exists?(config.ecs.cluster, family)
+        unless EcsManager.service_exists?(@target.cluster, family)
           exception "No service for #{family}! Please bootstrap or manually configure the service."
         end
         unless EcsManager.get_latest_task_definition_arn(family)
@@ -46,36 +46,36 @@ module Broadside
     end
 
     def bootstrap
-      if EcsManager.get_latest_task_definition_arn(family)
-        info("Task definition for #{family} already exists.")
-      else
-        unless @deploy_config.task_definition_config
+      unless EcsManager.get_latest_task_definition_arn(family)
+        unless @target.task_definition_config
           raise ArgumentError, "No first task definition and no :task_definition_config in '#{family}' configuration"
         end
 
         info "Creating an initial task definition for '#{family}' from the config..."
 
         EcsManager.ecs.register_task_definition(
-          @deploy_config.task_definition_config.merge(
+          @target.task_definition_config.merge(
             family: family,
             container_definitions: [DEFAULT_CONTAINER_DEFINITION.merge(container_definition)]
           )
         )
       end
 
-      if EcsManager.service_exists?(config.ecs.cluster, family)
+      run_commands(@target.bootstrap_commands)
+
+      if EcsManager.service_exists?(@target.cluster, family)
         info("Service for #{family} already exists.")
       else
-        unless @deploy_config.service_config
+        unless @target.service_config
           raise ArgumentError, "Service doesn't exist and no :service_config in '#{family}' configuration"
         end
 
         info "Service '#{family}' doesn't exist, creating..."
-        EcsManager.create_service(config.ecs.cluster, family, @deploy_config.service_config)
+        EcsManager.create_service(@target.cluster, family, @target.service_config)
       end
     end
 
-    def rollback(count = @deploy_config.rollback)
+    def rollback(count = @rollback)
       super do
         begin
           EcsManager.deregister_last_n_tasks_definitions(family, count)
@@ -95,32 +95,13 @@ module Broadside
 
     def run
       super do
-        update_task_revision
-
-        begin
-          run_command(@deploy_config.command)
-        ensure
-          EcsManager.deregister_last_n_tasks_definitions(family, 1)
-        end
-      end
-    end
-
-    # runs before deploy commands using the latest task definition
-    def run_predeploy
-      super do
-        update_task_revision
-
-        begin
-          @deploy_config.predeploy_commands.each { |command| run_command(command) }
-        ensure
-          EcsManager.deregister_last_n_tasks_definitions(family, 1)
-        end
+        run_commands(@command)
       end
     end
 
     def status
       super do
-        ips = EcsManager.get_running_instance_ips(config.ecs.cluster, family)
+        ips = EcsManager.get_running_instance_ips(@target.cluster, family)
         info "\n---------------",
           "\nDeployed task definition information:\n",
           Rainbow(PP.pp(EcsManager.get_latest_task_definition(family), '')).blue,
@@ -132,9 +113,7 @@ module Broadside
     end
 
     def container_definitions
-      super do
-        EcsManager.get_latest_task_definition(family)[:container_definitions]
-      end
+      EcsManager.get_latest_task_definition(family)[:container_definitions]
     end
 
     def logtail
@@ -142,7 +121,7 @@ module Broadside
         ip = get_running_instance_ip
         debug "Tailing logs for running container at ip #{ip}..."
         search_pattern = Shellwords.shellescape(family)
-        cmd = "docker logs -f --tail=#{@deploy_config.lines} `docker ps -n 1 --quiet --filter name=#{search_pattern}`"
+        cmd = "docker logs -f --tail=#{@lines} `docker ps -n 1 --quiet --filter name=#{search_pattern}`"
         tail_cmd = gen_ssh_cmd(ip) + " '#{cmd}'"
         exec tail_cmd
       end
@@ -170,7 +149,7 @@ module Broadside
     private
 
     def get_running_instance_ip
-      EcsManager.get_running_instance_ips(config.ecs.cluster, family).fetch(@deploy_config.instance)
+      EcsManager.get_running_instance_ips(@target.cluster, family).fetch(@target.instance)
     end
 
     # Creates a new task revision using current directory's env vars, provided tag, and configured options.
@@ -187,7 +166,7 @@ module Broadside
 
       # Deep merge doesn't work well with arrays (e.g. :container_definitions), so build the container first.
       updatable_container_definitions.first.merge!(container_definition)
-      revision.deep_merge!((@deploy_config.task_definition_config || {}).except(:container_definitions))
+      revision.deep_merge!((@target.task_definition_config || {}).except(:container_definitions))
 
       task_definition = EcsManager.ecs.register_task_definition(revision).task_definition
       debug "Successfully created #{task_definition.task_definition_arn}"
@@ -196,26 +175,26 @@ module Broadside
     # reloads the service using the latest task definition
     def update_service
       task_definition_arn = EcsManager.get_latest_task_definition_arn(family)
-      debug "Updating #{family} with scale=#{@deploy_config.scale} using task #{task_definition_arn}..."
+      debug "Updating #{family} with scale=#{@target.scale} using task #{task_definition_arn}..."
 
       update_service_response = EcsManager.ecs.update_service({
-        cluster: config.ecs.cluster,
-        desired_count: @deploy_config.scale,
+        cluster: @target.cluster,
+        desired_count: @target.scale,
         service: family,
         task_definition: task_definition_arn
-      }.deep_merge(@deploy_config.service_config || {}))
+      }.deep_merge(@target.service_config || {}))
 
       unless update_service_response.successful?
         exception('Failed to update service during deploy.', update_service_response.pretty_inspect)
       end
 
-      EcsManager.ecs.wait_until(:services_stable, { cluster: config.ecs.cluster, services: [family] }) do |w|
-        w.max_attempts = @deploy_config.timeout ? @deploy_config.timeout / config.ecs.poll_frequency : nil
+      EcsManager.ecs.wait_until(:services_stable, { cluster: @target.cluster, services: [family] }) do |w|
+        w.max_attempts = config.timeout ? config.timeout / config.ecs.poll_frequency : nil
         w.delay = config.ecs.poll_frequency
         seen_event = nil
 
         w.before_wait do |attempt, response|
-          debug "(#{attempt}/#{w.max_attempts}) Polling ECS for events..."
+          debug "(#{attempt}/#{w.max_attempts ? w.max_attempts : Float::INFINITY}) Polling ECS for events..."
           # skip first event since it doesn't apply to current request
           if response.services[0].events.first && response.services[0].events.first.id != seen_event && attempt > 1
             seen_event = response.services[0].events.first.id
@@ -225,36 +204,46 @@ module Broadside
       end
     end
 
-    def run_command(command)
-      command_name = command.join(' ')
-      run_task_response = EcsManager.run_task(config.ecs.cluster, family, command)
+    def run_commands(commands)
+      return if commands.nil? || commands.empty?
 
-      unless run_task_response.successful? && run_task_response.tasks.try(:[], 0)
-        exception("Failed to run #{command_name} task.", run_task_response.pretty_inspect)
-      end
+      update_task_revision
 
-      task_arn = run_task_response.tasks[0].task_arn
-      debug "Launched #{command_name} task #{task_arn}, waiting for completion..."
+      begin
+        Array.wrap(commands).each do |command|
+          command_name = command.join(' ')
+          run_task_response = EcsManager.run_task(@target.cluster, family, command)
 
-      EcsManager.ecs.wait_until(:tasks_stopped, { cluster: config.ecs.cluster, tasks: [task_arn] }) do |w|
-        w.max_attempts = nil
-        w.delay = config.ecs.poll_frequency
-        w.before_attempt do |attempt|
-          debug "Attempt #{attempt}: waiting for #{command_name} to complete..."
+          unless run_task_response.successful? && run_task_response.tasks.try(:[], 0)
+            exception("Failed to run #{command_name} task.", run_task_response.pretty_inspect)
+          end
+
+          task_arn = run_task_response.tasks[0].task_arn
+          debug "Launched #{command_name} task #{task_arn}, waiting for completion..."
+
+          EcsManager.ecs.wait_until(:tasks_stopped, { cluster: @target.cluster, tasks: [task_arn] }) do |w|
+            w.max_attempts = nil
+            w.delay = config.ecs.poll_frequency
+            w.before_attempt do |attempt|
+              debug "Attempt #{attempt}: waiting for #{command_name} to complete..."
+            end
+          end
+
+          info "#{command_name} task container logs:\n#{get_container_logs(task_arn)}"
+
+          if (code = EcsManager.get_task_exit_code(@target.cluster, task_arn, family)) == 0
+            debug "#{command_name} task #{task_arn} exited with status code 0"
+          else
+            exception "#{command_name} task #{task_arn} exited with a non-zero status code #{code}!"
+          end
         end
-      end
-
-      info "#{command_name} task container logs:\n#{get_container_logs(task_arn)}"
-
-      if (code = EcsManager.get_task_exit_code(config.ecs.cluster, task_arn, family)) == 0
-        debug "#{command_name} task #{task_arn} exited with status code 0"
-      else
-        exception "#{command_name} task #{task_arn} exited with a non-zero status code #{code}!"
+      ensure
+        EcsManager.deregister_last_n_tasks_definitions(family, 1)
       end
     end
 
     def get_container_logs(task_arn)
-      ip = EcsManager.get_running_instance_ips(config.ecs.cluster, family, task_arn).first
+      ip = EcsManager.get_running_instance_ips(@target.cluster, family, task_arn).first
       debug "Found ip of container instance: #{ip}"
 
       find_container_id_cmd = "#{gen_ssh_cmd(ip)} \"docker ps -aqf 'label=com.amazonaws.ecs.task-arn=#{task_arn}'\""
@@ -272,15 +261,16 @@ module Broadside
     end
 
     def container_definition
-      configured_containers = (@deploy_config.task_definition_config || {})[:container_definitions]
+      configured_containers = (@target.task_definition_config || {})[:container_definitions]
+
       if configured_containers && configured_containers.size > 1
         raise ArgumentError, 'Creating > 1 container definition not supported yet'
       end
 
       (configured_containers.try(:first) || {}).merge(
         name: family,
-        command: @deploy_config.command,
-        environment: @deploy_config.env_vars,
+        command: @command,
+        environment: @target.env_vars,
         image: image_tag
       )
     end