+++ lib/broadside/utils.rb
@@ -1,19 +1,19 @@
 module Broadside
   module Utils
     def debug(*args)
-      config.base.logger.debug(args.join(' '))
+      config.logger.debug(args.join(' '))
     end
 
     def info(*args)
-      config.base.logger.info(args.join(' '))
+      config.logger.info(args.join(' '))
     end
 
     def warn(*args)
-      config.base.logger.warn(args.join(' '))
+      config.logger.warn(args.join(' '))
     end
 
     def error(*args)
-      config.base.logger.error(args.join(' '))
+      config.logger.error(args.join(' '))
     end
 
     def exception(*args)