+++ spec/broadside/deploy/ecs_deploy_spec.rb
@@ -1,13 +1,11 @@
 require 'spec_helper'
 
 describe Broadside::EcsDeploy do
-  include_context 'full configuration'
+  include_context 'deploy configuration'
 
-  let(:app_name) { 'TEST_APP' }
-  let(:task_name) { 'TEST_TARGET' }
-  let(:family) { "#{app_name}_#{task_name}" }
-  let(:valid_options) { { target: task_name.to_sym } }
-  let(:deploy) { described_class.new(valid_options) }
+  let(:family) { "#{test_app}_#{test_target}" }
+  let(:target) { Broadside::Target.new(test_target, test_target_config) }
+  let(:deploy) { described_class.new(target, tag: 'tag_the_bag') }
 
   let(:api_request_log) { [] }
 
@@ -51,6 +49,31 @@ describe Broadside::EcsDeploy do
     }
   end
 
+  let(:arn) { 'arn:aws:ecs:us-east-1:1234' }
+  let(:existing_service) do
+    {
+      service_name: test_target.to_s,
+      service_arn: "#{arn}:service/#{test_target}",
+      deployments: [{ desired_count: 1, running_count: 1 }]
+    }
+  end
+  let(:stub_service_response) { { services: [existing_service], failures: [] } }
+  let(:task_definition_arn) { "#{arn}:task-definition/#{test_target}:1" }
+  let(:stub_task_definition_response) { { task_definition_arns: [task_definition_arn] } }
+  let(:stub_describe_task_definition_response) do
+    {
+      task_definition: {
+        task_definition_arn: task_definition_arn,
+        container_definitions: [
+          {
+            name: family
+          }
+        ],
+        family: family
+      }
+    }
+  end
+
   before(:each) { Broadside::EcsManager.instance_variable_set(:@ecs_client, ecs_stub) }
 
   it 'should instantiate an object' do
@@ -62,17 +85,42 @@ describe Broadside::EcsDeploy do
       expect { deploy.bootstrap }.to raise_error(//)
     end
 
-    it 'fails without service_config' do
-      deploy.deploy_config.task_definition_config = task_definition_config
+    context 'with an existing task definition' do
+      before(:each) do
+        ecs_stub.stub_responses(:list_task_definitions, stub_task_definition_response)
+        ecs_stub.stub_responses(:describe_task_definition, stub_describe_task_definition_response)
+      end
 
-      expect { deploy.bootstrap }.to raise_error(/Service doesn't exist and no :service_config/)
-    end
+      it 'fails without service_config' do
+        deploy.target.task_definition_config = task_definition_config
 
-    it 'succeeds' do
-      deploy.deploy_config.service_config = service_config
-      deploy.deploy_config.task_definition_config = task_definition_config
+        expect { deploy.bootstrap }.to raise_error(/Service doesn't exist and no :service_config/)
+      end
 
-      expect { deploy.bootstrap }.to_not raise_error
+      context 'with an existing service' do
+        before(:each) do
+          ecs_stub.stub_responses(:describe_services, stub_service_response)
+        end
+
+        it 'succeeds' do
+          deploy.target.service_config = service_config
+          deploy.target.task_definition_config = task_definition_config
+
+          expect { deploy.bootstrap }.to_not raise_error
+        end
+
+        context 'and some configured bootstrap commands' do
+          let(:commands) { [%w(foo bar baz)] }
+          let(:target) do
+            Broadside::Target.new(test_target, test_target_config.merge(bootstrap_commands: commands))
+          end
+
+          it 'runs bootstrap commands' do
+            expect(deploy).to receive(:run_commands).with(commands)
+            deploy.bootstrap
+          end
+        end
+      end
     end
   end
 
@@ -82,16 +130,6 @@ describe Broadside::EcsDeploy do
     end
 
     context 'with an existing service' do
-      let(:arn) { 'arn:aws:ecs:us-east-1:1234' }
-      let(:existing_service) do
-        {
-          service_name: task_name,
-          service_arn: "#{arn}:service/#{task_name}",
-          deployments: [{ desired_count: 1, running_count: 1 }]
-        }
-      end
-      let(:stub_service_response) { { services: [existing_service], failures: [] } }
-
       before(:each) do
         ecs_stub.stub_responses(:describe_services, stub_service_response)
       end
@@ -101,22 +139,6 @@ describe Broadside::EcsDeploy do
       end
 
       context 'with an existing task definition' do
-        let(:task_definition_arn) { "#{arn}:task-definition/#{task_name}:1" }
-        let(:stub_task_definition_response) { { task_definition_arns: [task_definition_arn] } }
-        let(:stub_describe_task_definition_response) do
-          {
-            task_definition: {
-              task_definition_arn: task_definition_arn,
-              container_definitions: [
-                {
-                  name: family
-                }
-              ],
-              family: family
-            }
-          }
-        end
-
         before(:each) do
           ecs_stub.stub_responses(:list_task_definitions, stub_task_definition_response)
           ecs_stub.stub_responses(:describe_task_definition, stub_describe_task_definition_response)
@@ -127,7 +149,7 @@ describe Broadside::EcsDeploy do
         end
 
         it 'should reconfigure the task definition' do
-          deploy.deploy_config.task_definition_config = task_definition_config
+          deploy.target.task_definition_config = task_definition_config
           deploy.short
 
           register_requests = api_request_log.select { |cmd| cmd.keys.first == :register_task_definition }
@@ -138,7 +160,7 @@ describe Broadside::EcsDeploy do
         end
 
         it 'should reconfigure the service definition' do
-          deploy.deploy_config.service_config = service_config
+          deploy.target.service_config = service_config
           deploy.short
 
           service_requests = api_request_log.select { |cmd| cmd.keys.first == :update_service }